Finished refactoring into Movables (not Recyclables). AI commented out (for now).

+++ src/main/java/edu/mines/csci598/recycler/frontend/graphics/Sprite.java
@@ -39,10 +39,10 @@ public class Sprite {
 
 
 
-    public void setY(double y){
+    public void setY(int y){
         location.setY(y);
     }
-    public void setX(double x){
+    public void setX(int x){
         location.setX(x);
     }
